- Fixes num_results miscount bug: - Parsers no longer have archive_contents class attribute - Consolidates modules to resolve circular deps
